+++ ft_str_printer.c
@@ -6,13 +6,12 @@
 /*   By: tamatsuu <tamatsuu@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/16 19:57:24 by tamatsuu          #+#    #+#             */
-/*   Updated: 2024/06/16 22:30:35 by tamatsuu         ###   ########.fr       */
+/*   Updated: 2024/06/18 02:17:31 by tamatsuu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libftprintf.h"
+#include "ft_printf.h"
 #include "libft/libft.h"
-#include <stdio.h>
 
 size_t	ft_print_str(t_format *fmt, char *str, int fd)
 {
@@ -41,7 +40,7 @@ size_t	ft_print_str_fd(t_format *fmt, char *s, int fd)
 		return (ft_putstr_null_fd(fmt, "(null)", fd));
 	if (fmt->dot)
 	{
-		while (i < fmt->precision)
+		while (i < (size_t)fmt->precision)
 			ft_putchar_fd(s[i++], fd);
 	}
 	else
@@ -59,7 +58,7 @@ size_t	ft_putstr_null_fd(t_format *fmt, char *nul_str, int fd)
 	i = 0;
 	if (fmt->dot)
 	{
-		while (i < fmt->precision)
+		while (i < (size_t)fmt->precision)
 			ft_putchar_fd(nul_str[i++], fd);
 	}
 	else
@@ -80,16 +79,16 @@ size_t	ft_putstr_mnw_str(t_format *fmt, char *s, int fd)
 	else
 		s_len = ft_strlen(s);
 	i = 0;
-	if (fmt->dot && fmt->precision < s_len)
+	if (fmt->dot && (size_t)fmt->precision < s_len)
 		s_len = fmt->precision;
 	if (fmt->flg & ZR_FLG)
 	{
-		while (i + s_len < fmt->min_w)
+		while (i + s_len < (size_t)fmt->min_w)
 			i = i + ft_putchar_fd_vp('0', fd);
 	}
 	else
 	{
-		while (i + s_len < fmt->min_w)
+		while (i + s_len < (size_t)fmt->min_w)
 			i = i + ft_putchar_fd_vp(' ', fd);
 	}
 	return (i);
